refactor: move MCP + analytics from Next.js to the gateway Worker
Enforces the hard rule: tools live in Cloudflare Workers, app lives in
Vercel, heavy processing lives in Fly.io.

Worker (backend/):
- backend/mcp/server.ts — MongoDB-backed, task-oriented MCP (ported from the
  Next.js version). Native mongodb driver via nodejs_compat + new MongoClient.
- backend/services/MongoService.ts — cached MongoClient connection helper
- backend/services/PublicAnalytics.ts — open-data analytics aggregation
- index.ts: /mcp now reads MongoDB; add GET /api/analytics; add
  /.well-known/oauth-authorization-server discovery; MONGODB_URI binding
- wrangler.jsonc: nodejs_compat flag + .well-known route
- package.json: add mongodb + jose

Auth: public MCP uses WorkOS OAuth (public PKCE client at identity.nyuchi.com),
not the M2M secret that the internal mongodb-mcp uses. JWKS verification only.

Removed from Next.js (violated the rule):
- src/lib/mcp/server.ts, src/app/api/mcp, src/app/api/analytics,
  src/app/.well-known/oauth-authorization-server
- analytics page now fetches the Worker's /api/analytics; MCP link → /mcp

.mcp.json → https://news.mukoko.com/mcp

Verified: backend + frontend typecheck clean; wrangler dry-run bundles the
mongodb driver (2.6 MB / 388 KB gzip).
