+++ backend/wrangler.jsonc
@@ -8,6 +8,8 @@
   "$schema": "node_modules/wrangler/config-schema.json",
   "name": "mukoko-news-gateway",
   "compatibility_date": "2025-10-01",
+  // nodejs_compat — required for the native MongoDB driver in the MCP/public API
+  "compatibility_flags": ["nodejs_compat"],
   "main": "./index.ts",
   "account_id": "125a2dfbc21f76a25c980609609e8218",
 
@@ -17,7 +19,8 @@
    */
   "routes": [
     { "pattern": "news.mukoko.com/mcp*", "zone_name": "mukoko.com" },
-    { "pattern": "news.mukoko.com/api/*", "zone_name": "mukoko.com" }
+    { "pattern": "news.mukoko.com/api/*", "zone_name": "mukoko.com" },
+    { "pattern": "news.mukoko.com/.well-known/oauth-authorization-server", "zone_name": "mukoko.com" }
   ],
   "vars": {
     "NODE_ENV": "production",
